+++ asdf_zarr/extensions.py
@@ -4,8 +4,8 @@ from .converter import ZarrConverter
 
 
 class ZarrExtension(asdf.extension.Extension):
-    extension_uri = "asdf://stsci.edu/example-project/tags/zarr-1.0.0"
-    tags = ["asdf://stsci.edu/example-project/tags/zarr-1.0.0"]
+    extension_uri = "asdf://asdf-format.org/zarr/tags/zarr-1.0.0"
+    tags = ["asdf://asdf-format.org/zarr/tags/zarr-1.0.0"]
     converters = [ZarrConverter()]