+++ CHANGES.rst
@@ -3,3 +3,4 @@
 
 - added changelog [#36]
 - pin zarr to < 3 until the 3.x API is stable [#43]
+- fix tags to use ``asdf-format.org`` instead of ``example-project`` [#44]